Remove compiler warnings

+++ agent-bootstrapper/test/com/thoughtworks/go/agent/bootstrapper/AgentBootstrapperTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 ThoughtWorks, Inc.
+ * Copyright 2017 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -207,7 +207,7 @@ public class AgentBootstrapperTest {
                                 incrementCount(descriptorContext);
                                 incrementCount(expectedContext);
                                 Integer expectedCount = (Integer) expectedContext.get(COUNT);
-                                assertThat((Integer) descriptorContext.get(COUNT), is(expectedCount));
+                                assertThat(descriptorContext.get(COUNT), is(expectedCount));
                                 if (expectedCount > 3) {
                                     ((AgentBootstrapper) descriptor.getBootstrapper()).stopLooping();
                                 }
